chore: prepare VIP-627 for proposal
